speech/pyproject.toml
[project]
name = "speech"
version = "0.1.0"
description = ""
authors = [
    {name = "kyeshmz",email = "kyeshimizu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper",
    "pyannote.audio",
    "torch",
    "torchaudio", # Often needed with torch for audio
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
